Restructured layout and added claims for bool and Locale as well as identity tokens for Apple and Google
