+++ Sources/JWTKit/Keys/JWK.swift
@@ -1,68 +1,3 @@
-/// A JSON Web Key Set.
-///
-/// A JSON object that represents a set of JWKs.
-/// Read specification (RFC 7517) https://tools.ietf.org/html/rfc7517.
-public struct JWKS: Decodable {
-    /// All JSON Web Keys
-    public var keys: [JWK]
-
-    /// Retrieves the desired key from the JSON Web Key Set
-    /// - Parameters:
-    ///   - identifier: The `kid` value to lookup.
-    ///   - type: The `kty` value.
-    public func find(identifier: String, type: JWK.KeyType) -> JWK? {
-        self.keys.filter { $0.keyType == type && $0.keyIdentifier?.string == identifier }.first
-    }
-
-    /// Retrieves the desired key from the JSON Web Key Set
-    /// - Parameters:
-    ///   - identifier: The `kid` value to lookup.
-    ///   - type: The `kty` value.
-    public func find(identifier: JWKIdentifier, type: JWK.KeyType) -> JWK? {
-        self.find(identifier: identifier.string, type: type)
-    }
-
-    /// Retrieves the desired keys from the JSON Web Key Set
-    /// Multiple keys can have the same `kid` if they have different `kty` values.
-    /// - Parameter identifier: The `kid` value to lookup.
-    public func find(identifier: JWKIdentifier) -> [JWK]? {
-        self.find(identifier: identifier.string)
-    }
-
-    /// Retrieves the desired keys from the JSON Web Key Set
-    /// Multiple keys can have the same `kid` if they have different `kty` values.
-    /// - Parameter identifier: The `kid` value to lookup.
-    public func find(identifier: String) -> [JWK]? {
-        self.keys.filter { $0.keyIdentifier?.string == identifier }
-    }
-}
-
-public struct JWKIdentifier: Hashable, Equatable {
-    public let string: String
-
-    public init(string: String) {
-        self.string = string
-    }
-}
-
-extension JWKIdentifier: Codable {
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        try self.init(string: container.decode(String.self))
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        try container.encode(self.string)
-    }
-}
-
-extension JWKIdentifier: ExpressibleByStringLiteral {
-    public init(stringLiteral value: String) {
-        self.init(string: value)
-    }
-}
-
 /// A JSON Web Key.
 ///
 /// Read specification (RFC 7517) https://tools.ietf.org/html/rfc7517.
