+++ Sources/JWTKit/Keys/JWKIdentifier.swift
@@ -1,0 +1,25 @@
+public struct JWKIdentifier: Hashable, Equatable {
+    public let string: String
+
+    public init(string: String) {
+        self.string = string
+    }
+}
+
+extension JWKIdentifier: Codable {
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+        try self.init(string: container.decode(String.self))
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try container.encode(self.string)
+    }
+}
+
+extension JWKIdentifier: ExpressibleByStringLiteral {
+    public init(stringLiteral value: String) {
+        self.init(string: value)
+    }
+}
