+++ Sources/JWTKit/Signing/ECDSA/ECDSAError.swift
@@ -1,0 +1,5 @@
+internal enum ECDSAError: Error {
+    case newKeyByCurveFailure
+    case generateKeyFailure
+    case signFailure
+}
