+++ Sources/JWTKit/Signing/ECDSA/ECDSAKey.swift
@@ -1,0 +1,41 @@
+import CJWTKitCrypto
+
+public final class ECDSAKey: OpenSSLKey {
+    public static func generate() throws -> ECDSAKey {
+        guard let c = EC_KEY_new_by_curve_name(NID_X9_62_prime256v1) else {
+            throw JWTError.signingAlgorithmFailure(ECDSAError.newKeyByCurveFailure)
+        }
+        guard EC_KEY_generate_key(c) != 0 else {
+            throw JWTError.signingAlgorithmFailure(ECDSAError.generateKeyFailure)
+        }
+        return .init(c)
+    }
+
+    public static func `public`<Data>(pem data: Data) throws -> ECDSAKey
+        where Data: DataProtocol
+    {
+        let c = try self.load(pem: data) { bio in
+            PEM_read_bio_EC_PUBKEY(convert(bio), nil, nil, nil)
+        }
+        return self.init(c)
+    }
+
+    public static func `private`<Data>(pem data: Data) throws -> ECDSAKey
+        where Data: DataProtocol
+    {
+        let c = try self.load(pem: data) { bio in
+            PEM_read_bio_ECPrivateKey(convert(bio), nil, nil, nil)
+        }
+        return self.init(c)
+    }
+
+    let c: OpaquePointer
+
+    init(_ c: OpaquePointer) {
+        self.c = c
+    }
+
+    deinit {
+        EC_KEY_free(self.c)
+    }
+}
