+++ Sources/JWTKit/Signing/ECDSA/ECDSASigner.swift
@@ -1,82 +1,6 @@
 import CJWTKitCrypto
 
-extension JWTSigner {
-    // MARK: ECDSA
-
-    public static func es256(key: ECDSAKey) -> JWTSigner {
-        return .init(algorithm: ECDSASigner(
-            key: key,
-            algorithm: convert(EVP_sha256()),
-            name: "ES256"
-        ))
-    }
-
-    public static func es384(key: ECDSAKey) -> JWTSigner {
-        return .init(algorithm: ECDSASigner(
-            key: key,
-            algorithm: convert(EVP_sha384()),
-            name: "ES384"
-        ))
-    }
-
-    public static func es512(key: ECDSAKey) -> JWTSigner {
-        return .init(algorithm: ECDSASigner(
-            key: key,
-            algorithm: convert(EVP_sha512()),
-            name: "ES512"
-        ))
-    }
-}
-
-public final class ECDSAKey: OpenSSLKey {
-    public static func generate() throws -> ECDSAKey {
-        guard let c = EC_KEY_new_by_curve_name(NID_X9_62_prime256v1) else {
-            throw JWTError.signingAlgorithmFailure(ECDSAError.newKeyByCurveFailure)
-        }
-        guard EC_KEY_generate_key(c) != 0 else {
-            throw JWTError.signingAlgorithmFailure(ECDSAError.generateKeyFailure)
-        }
-        return .init(c)
-    }
-
-    public static func `public`<Data>(pem data: Data) throws -> ECDSAKey
-        where Data: DataProtocol
-    {
-        let c = try self.load(pem: data) { bio in
-            PEM_read_bio_EC_PUBKEY(convert(bio), nil, nil, nil)
-        }
-        return self.init(c)
-    }
-
-    public static func `private`<Data>(pem data: Data) throws -> ECDSAKey
-        where Data: DataProtocol
-    {
-        let c = try self.load(pem: data) { bio in
-            PEM_read_bio_ECPrivateKey(convert(bio), nil, nil, nil)
-        }
-        return self.init(c)
-    }
-
-    let c: OpaquePointer
-
-    init(_ c: OpaquePointer) {
-        self.c = c
-    }
-
-    deinit {
-        EC_KEY_free(self.c)
-    }
-}
-
-// MARK: Private
-
-private enum ECDSAError: Error {
-    case newKeyByCurveFailure
-    case generateKeyFailure
-    case signFailure
-}
-
-private struct ECDSASigner: JWTAlgorithm, OpenSSLSigner {
+internal struct ECDSASigner: JWTAlgorithm, OpenSSLSigner {
     let key: ECDSAKey
     let algorithm: OpaquePointer
     let name: String
