+++ Sources/JWTKit/Signing/ECDSA/JWTSigner+ECDSA.swift
@@ -1,0 +1,27 @@
+import CJWTKitCrypto
+
+extension JWTSigner {
+    public static func es256(key: ECDSAKey) -> JWTSigner {
+        return .init(algorithm: ECDSASigner(
+            key: key,
+            algorithm: convert(EVP_sha256()),
+            name: "ES256"
+        ))
+    }
+
+    public static func es384(key: ECDSAKey) -> JWTSigner {
+        return .init(algorithm: ECDSASigner(
+            key: key,
+            algorithm: convert(EVP_sha384()),
+            name: "ES384"
+        ))
+    }
+
+    public static func es512(key: ECDSAKey) -> JWTSigner {
+        return .init(algorithm: ECDSASigner(
+            key: key,
+            algorithm: convert(EVP_sha512()),
+            name: "ES512"
+        ))
+    }
+}
