+++ Sources/JWTKit/Signing/HMAC/HMACError.swift
@@ -1,0 +1,5 @@
+internal enum HMACError: Error {
+    case initializationFailure
+    case updateFailure
+    case finalizationFailure
+}
