+++ Sources/JWTKit/Signing/RSA/RSAError.swift
@@ -1,0 +1,5 @@
+internal enum RSAError: Error {
+    case privateKeyRequired
+    case signFailure
+    case keyInitializationFailure
+}
